Generate queen promotions in qsearch
This was done to prevent illegal moves being generated in qsearch.
Before, only attacking moves were generated in qsearch, and all other
moves were ignored, including promotions. But this led to illegal moves
being generated during search if a pawn made an attacking move on its
opponets backrank.

The simple fix is to make sure at least one type of promotion move is
generated for pawn attacks that occur on the backrank.

Testing was briefly done to verify this needed change wasn't a
regression in strength.