Tweak static null move pruning to be more agressive
As we get closer to the horizon, the evaluation becomes a better
approximation for the state of the position, so we can lower the
margins we subtract from the static evaluation when we test if it
exceeds beta.

This tweak is also meant to largely replace futility pruning, since
it and static null move pruning are essentially doing the same task,
but from different perspectives. And thus the only way futility pruning
would be useful would be if it pruned moves that should be pruned, that
static null move pruning wouldn't prune.

With futility pruning, one case of moves that should be pruned would be
moves that are closer to the horizon that likely will fail low. This
would be done by using a smaller margin M, and doing the comparison
static_eval + M < alpha, to test if moves which are likely futile for
raising alpha (quiet moves, castling, etc.) can be skipped.

This test can equivalently be done once we make the move and change
perspectives, since static_eval + M < alpha => -(static_eval + M) >=
-alpha => static_eval' - M >= beta, where static_eval' = -static_eval
and is from the perspective of the opposite color after the move has
been made. Thus subtracting a smaller margin from the static eval of
the opponet in static null move pruning as the horizon approches, is
equivalent to adding a smaller margin to our static eval and seeing if
it's lower than alpha.

~26 Elo @ 10+0.1s
----------------------------------------------------------------------------------------
Score of Blunder 9.0.0-static-nmp-tweak vs Blunder 9.0.0: 1209 - 990 - 708  [0.538] 2907
...      Blunder 9.0.0-static-nmp-tweak playing White: 679 - 419 - 356  [0.589] 1454
...      Blunder 9.0.0-static-nmp-tweak playing Black: 530 - 571 - 352  [0.486] 1453
...      White vs Black: 1250 - 949 - 708  [0.552] 2907
Elo difference: 26.2 +/- 11.0, LOS: 100.0 %, DrawRatio: 24.4 %
SPRT: llr 2.96 (100.4%), lbound -2.94, ubound 2.94 - H1 was accepted